Integrate some ideas from #811
- Split the config object into three separate files based on content
- Cleanup main build.gradle.kts
